srv/requirements.txt
alembic==1.8.1
async-timeout==4.0.2
bcrypt==4.0.1
cachelib==0.9.0
click==8.1.3
cryptography==3.4.7
Deprecated==1.2.13
Flask==3.0.0
Flask-Bcrypt==1.0.1
Flask-Caching==1.10.1
Flask-Cors==3.0.10
Flask-JWT-Extended==4.6.0
Flask-Migrate==3.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==3.0.2
greenlet==1.1.3.post0
importlib-metadata==5.0.0
importlib-resources==5.10.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.3
MarkupSafe==2.1.1
marshmallow==3.18.0
opprint==1.0.3
packaging==21.3
PyJWT==2.6.0
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.21.0
PyYAML==6.0
pytest==7.1.3
pytest-cov==4.0.0
redis==4.3.4
six==1.16.0
SQLAlchemy==1.4.42
waitress==3.0.0
Werkzeug==3.0.0
wrapt==1.14.1
zipp==3.10.0
